Support for Checkpointing
Added support for checkpointing during training. While untested, this
should also allow for inferencing without needing to retraing the
network. A weird hack that was needed is that a warmup epoch is needed
before the saved checkpoint can be loaded. This is likely because the
lava dl package is doing something special the first time the forward
pass is called. On creation, the dimensions of the network are incorrect
and are only adjusted after a forward pass is called. For now, we simply
run a whole training epoch before loading a saved model, but it may be
possible to get away with just running a single mini-batch iteration. A
future TODO to be explored...