Support for tracking weights in axonal delays
Includes a plotting function that allows weights to be tracked on every
epoch of training. These waits are cloned, detached, and then migrated
to CPU after every training epoch, where they are stored until training
is done. After training completes, weights are passed to plotting
function that shows the updates as a heatmap is the switch
trackDelayWhileTraining is passed as a command line argument. This may
be useful for visualizing the activations of the layers during
inferencing and comparing the neuronal weights later on.