Support to Inspect Passive Denoising of Pinna
Applies parameterization and optimizations to the validate_si_snr script
to inspect look at the effect of a pinna on the entire validation set.
This was done to be able to condor to submit this across GPUs on condor,
as there are 43 subjects x 2 channels x 1250 speech orients x 1250 noise
orients to evaluate over the entire 60000 sample validation set.