Loading All CIPIC SOFA Files into Cipic Database
Previously, we only looked at a few of CIPIC subjects when looking to
add spatial cues to speech and noise. With this change, we all any model
in the cipic database to be used, even the two KEMAR subjects. These
have been hardcoded into the runtime database, since there are only 45
possible models.