Faster load from checkpoint
Because of the weirdness with the way axon delays are implemented, a warm up
epoch was implemented before loading a saved checkpoint. This had allowed the
network to re-configure the network dimesnions to reflect what the actual dims
of the weights should be. However, running a whole epoch was wasteful, so now
it only performs a single mini-batch iteration, which is a much lighter-weight
warm up.