Optimizing workflow to minimize wav file transfers
Originally, for both CIPIC and non CIPIC trails, all wav files (noisy, noise,
and clean) were loaded to CPU mem and transferred to GPU memory on every mini-
batch. However, noise is unused int he non-CIPIC trials and noisy is unused in
the CIPIC ones, so these files need not be used or even read. This change
allows those files to be ignored depending on what type of trial is being
launched.