Pitch predictor network with one hot encoding
Rather than try to have the network learn how to do a ISTFT, perform the
loss calculation in the frequency domain itself. This means that the
output out the network should be a one-hot encoded output that simply
selected the correct frequency bin that matches the detected pitch from
praat as closely as possible.