Fixes for validation loop pitch predictor network
Forgot to update the validation loop to the new loss function. Since we
are not output a one-hot vector to indicate the correct frequency bin of
the predicted pitch, we need to use categorical cross entropy loss
instead of MSE loss. Had done this for the training loop, but not for
the validation.