First attempt at CIPIC HRTF Predictor
Small CNN model to try and predict HRTF resposes from right ears from
the CIPIC database. Only 35 subjects from the CIPIC database have
anthropometric data collected, so we those subjects for the dataset.
Thus the dataset consists of 35*1250=43750 samples. Runs much faster
than the sdnn model, and seems to train well. Need to train for a while
and then plot results to compare to CIPIC responses to determine if it
is learning the correct things.