Minor Perf Improvements for Inference
Tried profiling the validate only model toidentify the bottleneck in
the app, but tools were very informative. These changes have minor code
improvements, but it seems the largest improvement will come from an app that
uses torch.compile on the model. However, the lab machine is old, and does
not have a CUDA version that will support torch.compile. Will attempt this on
the A100s, H100s, and L40s in the cluster GPUs.