Adding profiling and self-scheduling functionality
The profiler has shown some interesting results on our dedicated machine on
chechil. Now, we have improved the script to profile iterations when this
script is launched to CHTC. This should allow use to gain insights on what
hyper-parameters are better to use for the different GPU machines CHTC has to
offer.

Additionally, a self-scheduling functionality has been added. This is primarily
for usage on CHTC jobs. Often, jobs are given hard-time limits on job slots,
and this now allows the script to get as much compute out of the machine as the
time limit permits. In doing so, torch.compile can now be used on machines that
support it, which should improve runtimes. This also required the API to read
job attributes from a CHTC to be updated.