Beginning new CREPE Pitch Workflow
Previously, tensorflow was included in the docker image to allow the use
of CREPE as a dynamic pitch predictor for speech. However, using
tensorflow is just too slow (seems like it wasn't installed into the
docker correctly). Rather than fix the dependency issue, I simply wrote
the crepe predictions out to files, and removed tensorflow from the
docker image. This methodology will work so long as we don't change the
FFT parameters. If that needs to be changed at runtime, we'll need to
fix the tensor flow installation.