Changes to support pitch SNN training
Implements dataloaders for reading static crepe pitch predictions that were
output to a file. This approach is better as it's quicker and doesn't rely on
including tensorflow into the docker image. However, if we ever change the FFT
parameters, this will have to be redone.

For the pitch_snn file, seeds were used to ensure reproducability of
experiments. Also, the print/logging support was improved to be more robust
based on whether an experiment was running on a local machine or on a CHTC GPU
cluster. This should now be ready for sweeps.