Added support to help mitigate duplicated work
When collated_results.csv file is provided, this file is now checked
before the script runs to ensure that no time is spent computing
orientation pairs that had previously been computed. This requires the
collated_results.csv file on chechil and ResearchDrive to be kept
up-to-date as much as possible.